Rewrite to ts
